+++ src/index.ts
@@ -1,1 +1,2 @@
 export { withPlayground } from "@/decorators";
+export { generateAutocompletion } from "@/codemirror/extensions";